added buttons and switched to newtranstrad, site not really beatiful atm
