+++ main.py
@@ -1,31 +1,41 @@
 from flask import Flask, render_template, url_for, redirect, request, flash
-from modules.DNAGeneration import run
+from modules.TransTrad import autoData
 
 app = Flask(__name__)
 
-app.config.update( # used for flash() (no idea y)
+app.config.update(  # used for flash() (no idea y)
   TESTING=True,
   SECRET_KEY='192b9bdd22ab9ed4d12e236c78afcb9a393ec15f71bbf5dc987d54727823bcbf'
 )
 
-@app.route('/')
+
+@app.route('/', methods=['GET'])
 def home():
   return render_template('home.html')
 
 
 @app.route('/', methods=['POST'])
 def index_post():
-  text = request.form['text']
-  return redirect(url_for('results', text=text))
+  try:
+    DNA = request.form['DNA']
+    return redirect(url_for('results', input=DNA, type='DNA'))
+  except:
+    pass
+  try:
+    RNA = request.form['RNA']
+    return redirect(url_for('results', input=RNA, type='RNA'))
+  except:
+    pass
 
 
 @app.route('/results')
 def results():
-  text = request.args['text']
-  data = run(text)
+  text = request.args['input']
+  type = request.args['type']
+  data = autoData(text, type)
   if data is None:
     print('The text did not work')
-    flash('L\'ADN que vous avez rentré est incorrecte...'
+    flash('Vous n\'avez pas rentré une valeur correcte...'
           )  # Est affiché dans home.html
     return redirect(url_for('home'))
   else:
