+++ notes/alex_suivi.doc
@@ -1,3 +1,6 @@
 la ligne de code : ARN_start = 'AUG' + ARN.split('AUG', 1)[1] a une IndexError quand 'AUG' n'est pas générer par gen_adn() (logique)
 au debut j'ai fait un try: except mais faut regenerer gen_adn donc boucle infini (d'erreurs) quand c'est faux, donc j'avais pensé a faire une boucle while qui englobe les deux fonctions mais c'est vraiment degueulasse. En attendant je fais juste une boucle conditionnelle if 'AUG' in ARN_start: ARN_start = 'AUG' + ARN.split('AUG', 1)[1] else : sys.exit()
 J'ai au final réussi en gardant ma boucle conditionnelle et je retourne False si il n'y a pas de codon d'initiation et dans main.py je verifie si transcription() retourne False, si c'est le cas je relance le programme pour regénérer un nouveau brin d'ADN
+
+j'ai pas mal suer sur le css et l'html notamment pour centrer le brin d'adn dans chaque div dans results.html, comme generationdata est une variable, le div container ne centre pas le brin d'adn car il ne connait pas sa longueur j'ai donc defini width d'une certaine longueur qui fait que le brin d'adn ne depasse pas à <= 100 caracteres
+egalement sur les outlines des deux boutons principaux qui bugais mais facilement rectifiable en rajoutant la balise <span> et le css correspondant (span {...})
