+++ templates/results.html
@@ -18,18 +18,13 @@
     </div>
     
     <div class="container" style="border:thick double #ff7024;width:1200px;">
-      <h1 class="display-6 fw-semibold" style="font-size: 25px;">ARNm_START :</h1>
+       <h1 class="display-6 fw-semibold" style="font-size: 25px;">Protéine(s) :</h1>
       <p>{{ generationdata[2] }}</p>
-    </div>
-    
-    <div class="container" style="border:thick double #ff7024;width:1200px;">
-       <h1 class="display-6 fw-semibold" style="font-size: 25px;">Protéine :</h1>
-      <p>{{ generationdata[3] }}</p>
     </div>
     <br>
 
 </div>
 <div class="menu">
-     <a class="rand-dna custom-btn btn-7" type="button" href="/results?text=None"><span>Regenerer brin d'ADN</span></a></div>
+     <a class="rand-dna custom-btn btn-7" type="button" href="/results?input=None&type=DNA"><span>Regenerer brin d'ADN</span></a></div>
 </div>
 {% endblock %}
